requirements.txt
requests
PyYaml>=5.4.1
fire==0.3.1
attrs==21.4.0
pyattck-data-models>=1.0.1